+++ cardano-crypto-class/testlib/Test/Crypto/DSIGN.hs
@@ -38,7 +38,6 @@ import Cardano.Crypto.Libsodium
 import Text.Show.Pretty (ppShow)
 
 #ifdef SECP256K1_ENABLED
-import Control.Monad (replicateM)
 import qualified GHC.Exts as GHC
 #endif
 
@@ -48,6 +47,7 @@ import Data.Proxy (Proxy (..))
 import Data.Maybe (fromJust)
 
 import Control.Exception (evaluate, bracket)
+import Control.Monad (replicateM)
 
 import Cardano.Crypto.DSIGN (
   MockDSIGN,
@@ -91,17 +91,23 @@ import Cardano.Crypto.DSIGN (
   genKeyDSIGNM,
 
   getSeedDSIGNM,
-  forgetSignKeyDSIGNM
+  forgetSignKeyDSIGNM,
+  BLS12381CurveConstraints,
+  BLS12381DSIGN,
+
+  DSIGNAggregatable (..),
+  sizeProofOfPossessionDSIGN,
+  encodeProofOfPossessionDSIGN,
+  decodeProofOfPossessionDSIGN
   )
 import Cardano.Binary (FromCBOR, ToCBOR)
+import Cardano.Crypto.EllipticCurve.BLS12_381 (Curve1, Curve2)
 import Cardano.Crypto.PinnedSizedBytes (PinnedSizedBytes)
 import Cardano.Crypto.DirectSerialise
 import Test.Crypto.Util (
-#if SECP256K1_ENABLED
   BadInputFor,
   genBadInputFor,
   shrinkBadInputFor,
-#endif
   Message,
   prop_raw_serialise,
   prop_raw_deserialise,
@@ -119,6 +125,7 @@ import Test.Crypto.Util (
   directSerialiseToBS,
   directDeserialiseFromBS,
   hexBS,
+  BLS12381SignContext (..),
   )
 import Cardano.Crypto.Libsodium.MLockedSeed
 
@@ -149,6 +156,12 @@ ed25519SigGen = defaultSigGen
 
 ed448SigGen :: Gen (SigDSIGN Ed448DSIGN)
 ed448SigGen = defaultSigGen
+
+blsSigGen :: forall curve. BLS12381CurveConstraints curve => Gen (SigDSIGN (BLS12381DSIGN curve))
+blsSigGen = do
+  msg :: Message <- arbitrary
+  BLS12381SignContext dst aug :: BLS12381SignContext <- arbitrary
+  signDSIGN (dst,aug) msg <$> defaultSignKeyGen @(BLS12381DSIGN curve)
 
 #ifdef SECP256K1_ENABLED
 ecdsaSigGen :: Gen (SigDSIGN EcdsaSecp256k1DSIGN)
@@ -181,6 +194,22 @@ defaultSigGen = do
   msg :: Message <- arbitrary
   signDSIGN () msg <$> defaultSignKeyGen
 
+defaultProofOfPossessionGen
+  :: forall v.
+     ( DSIGNAggregatable v
+     )
+  => Gen (ContextDSIGN v)
+  -> Gen (ProofOfPossessionDSIGN v)
+defaultProofOfPossessionGen genContext = do
+  ctx <- genContext
+  sk  <- defaultSignKeyGen @v
+  pure $ proveProofOfPossessionDSIGN ctx sk
+
+blsContextGen :: Gen (ContextDSIGN (BLS12381DSIGN curve))
+blsContextGen = do
+  BLS12381SignContext dst aug <- arbitrary
+  pure (dst, aug)
+
 -- Used for adjusting no of quick check tests
 -- By default up to 100 tests are performed which may not be enough to catch hidden bugs
 testEnough :: Spec -> Spec
@@ -199,6 +228,8 @@ tests lock =
        testDSIGNAlgorithm mockSigGen (arbitrary @Message) "MockDSIGN"
        testDSIGNAlgorithm ed25519SigGen (arbitrary @Message) "Ed25519DSIGN"
        testDSIGNAlgorithm ed448SigGen (arbitrary @Message) "Ed448DSIGN"
+       testDSIGNAlgorithmWithContext (blsContextGen @Curve1) (blsSigGen @Curve1) (arbitrary @Message) "BLS12381MinVerKeyDSIGN"
+       testDSIGNAlgorithmWithContext (blsContextGen @Curve2) (blsSigGen @Curve2) (arbitrary @Message) "BLS12381MinSigDSIGN"
 #ifdef SECP256K1_ENABLED
        testDSIGNAlgorithm ecdsaSigGen genEcdsaMsg "EcdsaSecp256k1DSIGN"
        testDSIGNAlgorithm schnorrSigGen (arbitrary @Message) "SchnorrSecp256k1DSIGN"
@@ -211,12 +242,15 @@ tests lock =
 #endif
      describe "MLocked" $ do
       testDSIGNMAlgorithm lock (Proxy @Ed25519DSIGN) "Ed25519DSIGN"
+     describe "Aggregatable" $ do
+      testDSIGNAggregatableWithContext (Proxy @(BLS12381DSIGN Curve1)) (blsContextGen @Curve1) (arbitrary @Message) "BLS12381MinVerKeyDSIGN"
+      testDSIGNAggregatableWithContext (Proxy @(BLS12381DSIGN Curve2)) (blsContextGen @Curve2) (arbitrary @Message) "BLS12381MinSigDSIGN"
 
-testDSIGNAlgorithm :: forall (v :: Type) (a :: Type).
+testDSIGNAlgorithmWithContext :: forall (v :: Type) (a :: Type).
   (DSIGNAlgorithm v,
    Signable v a,
-   ContextDSIGN v ~ (),
    Show a,
+   Show (ContextDSIGN v),
    Eq (SignKeyDSIGN v),
    Eq a,
    ToCBOR (VerKeyDSIGN v),
@@ -225,11 +259,12 @@ testDSIGNAlgorithm :: forall (v :: Type) (a :: Type).
    FromCBOR (SignKeyDSIGN v),
    ToCBOR (SigDSIGN v),
    FromCBOR (SigDSIGN v)) =>
+  Gen (ContextDSIGN v) ->
   Gen (SigDSIGN v) ->
   Gen a ->
   String ->
   Spec
-testDSIGNAlgorithm genSig genMsg name = testEnough . describe name $ do
+testDSIGNAlgorithmWithContext genContext genSig genMsg name = testEnough . describe name $ do
   describe "serialization" $ do
     describe "raw" $ do
       prop "VerKey serialization" .
@@ -296,7 +331,7 @@ testDSIGNAlgorithm genSig genMsg name = testEnough . describe name $ do
         prop_cbor_direct_vs_class encodeSigDSIGN
     describe "verify" $ do
       prop "signing and verifying with matching keys" .
-        forAllShow ((,) <$> genMsg <*> defaultSignKeyGen @v) ppShow $
+        forAllShow ((,,) <$> genContext <*> genMsg <*> defaultSignKeyGen @v) ppShow $
         prop_dsign_verify
       prop "verifying with wrong key" .
         forAllShow genWrongKey ppShow $
@@ -313,18 +348,40 @@ testDSIGNAlgorithm genSig genMsg name = testEnough . describe name $ do
       prop "VerKey rawDeserialise" . forAllShow (defaultVerKeyGen @v) ppShow $ \vk ->
         prop_no_thunks (fromJust $! rawDeserialiseVerKeyDSIGN @v . rawSerialiseVerKeyDSIGN $ vk)
   where
-    genWrongKey :: Gen (a, SignKeyDSIGN v, SignKeyDSIGN v)
+    genWrongKey :: Gen (ContextDSIGN v, a, SignKeyDSIGN v, SignKeyDSIGN v)
     genWrongKey = do
+      ctx <- genContext
       sk1 <- defaultSignKeyGen
       sk2 <- Gen.suchThat defaultSignKeyGen (/= sk1)
       msg <- genMsg
-      pure (msg, sk1, sk2)
-    genWrongMsg :: Gen (a, a, SignKeyDSIGN v)
+      pure (ctx, msg, sk1, sk2)
+    genWrongMsg :: Gen (ContextDSIGN v, a, a, SignKeyDSIGN v)
     genWrongMsg = do
+      ctx <- genContext
       msg1 <- genMsg
       msg2 <- Gen.suchThat genMsg (/= msg1)
       sk <- defaultSignKeyGen
-      pure (msg1, msg2, sk)
+      pure (ctx, msg1, msg2, sk)
+
+testDSIGNAlgorithm :: forall v a.
+  ( DSIGNAlgorithm v
+  , Signable v a
+  , ContextDSIGN v ~ ()
+  , Show a
+  , Eq (SignKeyDSIGN v)
+  , Eq a
+  , ToCBOR (VerKeyDSIGN v)
+  , FromCBOR (VerKeyDSIGN v)
+  , ToCBOR (SignKeyDSIGN v)
+  , FromCBOR (SignKeyDSIGN v)
+  , ToCBOR (SigDSIGN v)
+  , FromCBOR (SigDSIGN v)
+  ) =>
+  Gen (SigDSIGN v) ->
+  Gen a ->
+  String ->
+  Spec
+testDSIGNAlgorithm = testDSIGNAlgorithmWithContext (pure ())
 
 testDSIGNMAlgorithm
   :: forall v. ( -- change back to DSIGNMAlgorithm when unsound API is phased out
@@ -571,30 +628,28 @@ prop_dsignm_seed_roundtrip p seedPSB = ioProperty . withMLockedSeedFromPSB seedP
 prop_dsign_verify
   :: forall (v :: Type) (a :: Type) .
      ( DSIGNAlgorithm v
-     , ContextDSIGN v ~ ()
      , Signable v a
      )
-  => (a, SignKeyDSIGN v)
+  => (ContextDSIGN v,a, SignKeyDSIGN v)
   -> Property
-prop_dsign_verify (msg, sk) =
-  let signed = signDSIGN () msg sk
+prop_dsign_verify (ctx, msg, sk) =
+  let signed = signDSIGN ctx msg sk
       vk = deriveVerKeyDSIGN sk
-    in verifyDSIGN () vk msg signed === Right ()
+    in verifyDSIGN ctx vk msg signed === Right ()
 
 -- If we sign a message with one key, and try to verify with another, then
 -- verification fails.
 prop_dsign_verify_wrong_key
   :: forall (v :: Type) (a :: Type) .
      ( DSIGNAlgorithm v
-     , ContextDSIGN v ~ ()
      , Signable v a
      )
-  => (a, SignKeyDSIGN v, SignKeyDSIGN v)
+  => (ContextDSIGN v, a, SignKeyDSIGN v, SignKeyDSIGN v)
   -> Property
-prop_dsign_verify_wrong_key (msg, sk, sk') =
-  let signed = signDSIGN () msg sk
+prop_dsign_verify_wrong_key (ctx, msg, sk, sk') =
+  let signed = signDSIGN ctx msg sk
       vk' = deriveVerKeyDSIGN sk'
-    in verifyDSIGN () vk' msg signed =/= Right ()
+    in verifyDSIGN ctx vk' msg signed =/= Right ()
 
 prop_dsignm_verify_pos
   :: forall v. (DSIGNMAlgorithm v, ContextDSIGN v ~ (), Signable v Message)
@@ -632,13 +687,13 @@ prop_dsignm_verify_neg_key lock _ msg seedPSB seedPSB' =
 -- message, then verification fails.
 prop_dsign_verify_wrong_msg
   :: forall (v :: Type) (a :: Type) .
-  (DSIGNAlgorithm v, Signable v a, ContextDSIGN v ~ ())
-  => (a, a, SignKeyDSIGN v)
+  (DSIGNAlgorithm v, Signable v a)
+  => (ContextDSIGN v ,a, a, SignKeyDSIGN v)
   -> Property
-prop_dsign_verify_wrong_msg (msg, msg', sk) =
-  let signed = signDSIGN () msg sk
+prop_dsign_verify_wrong_msg (ctx, msg, msg', sk) =
+  let signed = signDSIGN ctx msg sk
       vk = deriveVerKeyDSIGN sk
-    in verifyDSIGN () vk msg' signed =/= Right ()
+    in verifyDSIGN ctx vk msg' signed =/= Right ()
 
 #ifdef SECP256K1_ENABLED
 instance Arbitrary (BadInputFor MessageHash) where
@@ -657,11 +712,14 @@ testEcdsaWithHashAlgorithm ::
   Proxy h -> String -> Spec
 testEcdsaWithHashAlgorithm _ name = testEnough . describe name $ do
   prop "Ecdsa sign and verify" .
-    forAllShow ((,) <$> genMsg <*> defaultSignKeyGen @EcdsaSecp256k1DSIGN) ppShow $
+    forAllShow ((,,) <$> genContext <*> genMsg <*> defaultSignKeyGen @EcdsaSecp256k1DSIGN) ppShow $
       prop_dsign_verify
   where
     genMsg :: Gen MessageHash
     genMsg = hashAndPack (Proxy @h) . messageBytes <$> arbitrary
+    genContext :: Gen (ContextDSIGN EcdsaSecp256k1DSIGN)
+    genContext = pure ()
+
 #endif
 
 prop_dsignm_verify_neg_msg
@@ -680,3 +738,114 @@ prop_dsignm_verify_neg_msg lock _ a a' =
       a /= a' ==> verifyDSIGN () vk a' sig =/= Right ()
 
 -- TODO: verify that DSIGN and DSIGNM implementations match (see #363)
+
+-- Tests and instances for DSIGNAggregatable v
+
+instance DSIGNAggregatable v => Arbitrary (BadInputFor (ProofOfPossessionDSIGN v)) where
+  arbitrary =
+    genBadInputFor (fromIntegral $ sizeProofOfPossessionDSIGN (Proxy @v))
+  shrink = shrinkBadInputFor
+
+testDSIGNAggregatableWithContext
+  :: forall (v :: Type).
+     ( DSIGNAggregatable v
+     , Signable v Message
+     , Show (ContextDSIGN v)
+     , ToCBOR (ProofOfPossessionDSIGN v)
+     , FromCBOR (ProofOfPossessionDSIGN v)
+     )
+  => Proxy v
+  -> Gen (ContextDSIGN v)
+  -> Gen Message
+  -> String
+  -> Spec
+testDSIGNAggregatableWithContext _ genContext genMsg name = testEnough . describe name $ do
+  describe "serialization" $ do
+    describe "raw" $ do
+      prop "PoP serialization" .
+        forAllShow (defaultProofOfPossessionGen @v genContext)
+                   ppShow $
+                   prop_raw_serialise rawSerialiseProofOfPossessionDSIGN rawDeserialiseProofOfPossessionDSIGN
+      prop "PoP deserialization (wrong length)" $ prop_raw_deserialise (rawDeserialiseProofOfPossessionDSIGN @v)
+      prop "PoP fail fromCBOR" $ prop_bad_cbor_bytes @(ProofOfPossessionDSIGN v)
+    describe "size" $ do
+      prop "PoP" .
+        forAllShow (defaultProofOfPossessionGen @v genContext)
+                   ppShow $
+                   prop_size_serialise rawSerialiseProofOfPossessionDSIGN (sizeProofOfPossessionDSIGN (Proxy @v))
+    describe "direct CBOR" $ do
+      prop "PoP" .
+        forAllShow (defaultProofOfPossessionGen @v genContext)
+                   ppShow $
+                   prop_cbor_with encodeProofOfPossessionDSIGN decodeProofOfPossessionDSIGN
+    describe "To/FromCBOR class" $ do
+      prop "PoP" . forAllShow (defaultProofOfPossessionGen @v genContext) ppShow $ prop_cbor
+    describe "ToCBOR size" $ do
+      prop "PoP" . forAllShow (defaultProofOfPossessionGen @v genContext) ppShow $ prop_cbor_size
+    describe "direct matches class" $ do
+      prop "PoP" .
+        forAllShow (defaultProofOfPossessionGen @v genContext) ppShow $
+        prop_cbor_direct_vs_class encodeProofOfPossessionDSIGN
+  describe "aggregate" $ do
+    prop "aggregate verify positive" $
+      forAllShow (genAggregateCase genContext genMsg) ppShow $
+          \(ctx, msg, vksPops, sigs) ->
+            case aggregateSigDSIGN @v sigs of
+              Left _  -> counterexample "aggregateSigDSIGN failed" False
+              Right s -> verifyAggregateDSIGN @v ctx vksPops msg s === Right ()
+    prop "aggregate verify negative (wrong message)" $
+      forAllShow (genAggregateCase2Msgs genContext genMsg) ppShow $
+        \(ctx, msg, msg', vksPops, sigs) ->
+          msg /= msg' ==> case aggregateSigDSIGN @v sigs of
+            Left _  -> counterexample "aggregateSigDSIGN failed" False
+            Right s -> verifyAggregateDSIGN @v ctx vksPops msg' s =/= Right ()
+    prop "aggregate verify negative (wrong PoP)" $
+        forAllShow (genAggregateCaseAtLeast2 genContext genMsg) ppShow $
+          \(ctx, msg, vksPops, sigs) ->
+            case vksPops of
+              (a:b:rest) ->
+                let vksPops' = (fst a, snd b) : (fst b, snd a) : rest
+                in case aggregateSigDSIGN @v sigs of
+                    Left _  -> counterexample "aggregateSigDSIGN failed" False
+                    Right s -> verifyAggregateDSIGN @v ctx vksPops' msg s =/= Right ()
+              _ ->
+                counterexample "genAggregateCaseAtLeast2 produced <2 entries (bug in generator)" False
+    describe "NoThunks" $ do
+      prop "PoP" . forAllShow (defaultProofOfPossessionGen @v genContext) ppShow $ prop_no_thunks
+      prop "PoP rawSerialise" . forAllShow (defaultProofOfPossessionGen @v genContext) ppShow $ \pop ->
+        prop_no_thunks (rawSerialiseProofOfPossessionDSIGN pop)
+      prop "PoP rawDeserialise" . forAllShow (defaultProofOfPossessionGen @v genContext) ppShow $ \pop ->
+        prop_no_thunks (fromJust $! rawDeserialiseProofOfPossessionDSIGN @v . rawSerialiseProofOfPossessionDSIGN $ pop)
+  where
+    genAggregateCase genCtx genMsg' = do
+      ctx <- genCtx
+      msg <- genMsg'
+      -- These crypto operations can be expensive, so limit the number of
+      -- signatures to a reasonable number for testing.
+      n   <- Gen.chooseInt (1, 8)
+      sks <- replicateM n (defaultSignKeyGen @v)
+      let vksPops = [ ( deriveVerKeyDSIGN sk
+                     , proveProofOfPossessionDSIGN ctx sk
+                     )
+                   | sk <- sks
+                   ]
+          sigs = [ signDSIGN ctx msg sk | sk <- sks ]
+      pure (ctx, msg, vksPops, sigs)
+    genAggregateCase2Msgs genCtx genMsg' = do
+      (ctx, msg, vksPops, sigs) <- genAggregateCase genCtx genMsg'
+      msg' <- Gen.suchThat genMsg' (/= msg)
+      pure (ctx, msg, msg', vksPops, sigs)
+    genAggregateCaseAtLeast2 genCtx genMsg' = do
+      ctx <- genCtx
+      msg <- genMsg'
+      -- These crypto operations can be expensive, so limit the number of
+      -- signatures to a reasonable number for testing.
+      n   <- Gen.chooseInt (2, 8)
+      sks <- replicateM n (defaultSignKeyGen @v)
+      let vksPops = [ ( deriveVerKeyDSIGN sk
+                     , proveProofOfPossessionDSIGN ctx sk
+                     )
+                   | sk <- sks
+                   ]
+          sigs = [ signDSIGN ctx msg sk | sk <- sks ]
+      pure (ctx, msg, vksPops, sigs)